Extract express app setup to config dir
In order to support the continued development of the application,
the logic for fetching the correct environment configuration and
setting up the express app on the server has been extracted to a
new config directory.

As new target environments are added, will add the correct information
to each environment-specific file in the config/env directory.