move spotify api interactions to service modules
The previous method of interacting with the spotify API was to create
an event emitter object that would shoot off a "done" event when data
was returned from the Spotify public API.

Now, the individual Spotify API calls are extracted into separate
functions that are defined using a middleware pattern, that allows them
to be composed into middleware stacks within the router. These
middleware implementations work by adding the Spotify API info to the
express request object, which is then passed down the stack.

At this point, the entire previous functionality of the related artist
search demo has been transitioned over to a react front-end app and a
more modular and reconfigurable express app on the back-end.

Closes #8